rework test suite to install franchise in local directory. Ignore-this: 679c5e77416d3b4b32bbcaa2261980f5 With this change we redefine $HOME to be a local directory, which means we can run ghc-pkg register without overwriting the user's actual package registry.
Note: this probably doesn't yet work on windows.  I don't know how to
redefine the location of the user package registry on windows.  But
I think this is the right way to do things.  The test suite must
install haskell packages, and the best way to do this is by installing
them in the test directory.

Also note that I've added a few very ugly exports that *will* be
cleaned up, once I've figured out a half-way decent API for things
like installing packages to a local directory.